+++ lib/src/models/models.dart
@@ -1,3 +1,4 @@
+export 'enum.dart';
 export 'action/action.dart';
 export 'action/action_result.dart';
 export 'auth/auth.dto.dart';